Revert "chore: codemirror-revert"
