+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@devtron-labs/devtron-fe-common-lib",
-    "version": "1.7.0-patch-1",
+    "version": "1.7.0",
     "description": "Supporting common component library",
     "type": "module",
     "main": "dist/index.js",
@@ -64,8 +64,6 @@
         "json-schema": "^0.4.0",
         "lint-staged": "^12.5.0",
         "moment": "^2.29.4",
-        "monaco-editor": "0.44.0",
-        "monaco-yaml": "5.1.1",
         "prettier": "^3.1.1",
         "react-ga4": "^1.4.1",
         "react-toastify": "9.1.3",
@@ -117,7 +115,6 @@
         "react-codemirror-merge": "4.23.6",
         "react-dates": "^21.8.0",
         "react-diff-viewer-continued": "^3.4.0",
-        "react-monaco-editor": "^0.54.0",
         "react-virtualized-sticky-tree": "^3.0.0-beta18",
         "sass": "^1.69.7",
         "tslib": "2.7.0",
@@ -128,9 +125,6 @@
         "react-dates": {
             "react": "^17.0.2",
             "react-dom": "^17.0.2"
-        },
-        "react-monaco-editor": {
-            "monaco-editor": "0.44.0"
         },
         "vite-plugin-svgr": {
             "vite": "5.4.14"
