+++ src/Common/CodeEditor/Extensions/findAndReplace.tsx
