+++ src/Common/CodeEditor/Extensions/readOnlyTooltip.ts
