+++ src/Common/CodeEditor/index.ts
@@ -15,4 +15,4 @@
  */
 
 export { default as CodeEditor } from './CodeEditor'
-export * from './types'
+export type { CodeEditorProps } from './types'
